Add example for stm32f30x
This patch adds an example using the stm32f30x HAL.  It was tested on a
STM32F303 on a STM32F3 Discovery board.  It includes the OpenOCD
configuration and the GDB script required to flash the program on the
MCU.  The Cargo project is based on the cortex-m-quickstart project by
the Cortex-M team.